Many features implementations and improvements

+++ src/RentBike.API/RentBike.Application/Handlers/Commands/DeleteBikeCommandHandler.cs
@@ -11,18 +11,23 @@ namespace RentBike.Application.Handlers.Commands
         readonly ILogger<DeleteBikeCommandHandler> _logger;
         readonly IAdminUserRepository _adminUserRepository;
         readonly IBikeRepository _bikeRepository;
+        readonly IRentRepository _rentRepository;
 
         public DeleteBikeCommandHandler(ILogger<DeleteBikeCommandHandler> logger,
-            IAdminUserRepository adminUserRepository, IBikeRepository bikeRepository)
+            IAdminUserRepository adminUserRepository, IBikeRepository bikeRepository
+            ,IRentRepository rentRepository)
         {
             _logger = logger;
             _adminUserRepository = adminUserRepository;
             _bikeRepository = bikeRepository;
+            _rentRepository = rentRepository;
         }
         public async Task Handle(DeleteBikeCommand request, CancellationToken cancellationToken)
         {
             var adminUser = await _adminUserRepository.GetById(Guid.Parse(request.AdminUserId)) ?? throw new AdminUserNotFoundException();
             var bike = await _bikeRepository.GetById(request.Id) ?? throw new BikeNotFoundException();
+            var hasRent = (await _rentRepository.Find(p => p.BikeId == bike.Id)).Any();
+            if (hasRent) throw new RemoveBikeException();
             await _bikeRepository.Remove(bike);
         }
     }
